# docs/openapi.yaml
openapi: 3.1.0
info:
  title: TechNest Shop API
  version: 1.0.0
  description: >
    Versioned OpenAPI contract for the current TechNest Shop backend. This file is
    the authoritative machine-readable API contract; README keeps the operational
    snapshot and docs/API_CONTRACT.md keeps a concise human summary.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /
...
paths:
...
  /api/products:
    get:
      tags: [Catalog]
      operationId: listProducts
      summary: List products
      security: []
      parameters:
        - in: query
          name: categoryId
          schema:
            type: integer
            format: int64
            minimum: 1
        - in: query
          name: cat
          schema:
            type: string
          description: Category name alias, for example `laptop`, `phone`, `all`.
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: string
          description: Decimal string.
        - in: query
          name: maxPrice
          schema:
            type: string
          description: Decimal string.
        - in: query
          name: brand
          schema:
            type: string
          description: Explicit brand name, case-insensitive.
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      tags: [Catalog]
      operationId: createProduct
      summary: Create product
      description: Staff or admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpsertRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/products/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Catalog]
      operationId: getProduct
      summary: Get product by id
      security: []
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Catalog]
      operationId: updateProduct
      summary: Update product
      description: Staff or admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpsertRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Catalog]
      operationId: deleteProduct
      summary: Delete product
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Product deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/reviews/product/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductIdPath'
    get:
      tags: [Reviews]
      operationId: listProductReviews
      summary: List approved reviews for a product
      security: []
      responses:
        '200':
          description: Approved reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
    post:
      tags: [Reviews]
      operationId: createReview
      summary: Create review for a product
      description: >
        Authenticated user only. Business logic only allows customers to review
        products from delivered and paid orders. New reviews start as pending.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '200':
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/reviews/pending-count:
    get:
      tags: [Reviews]
      operationId: getPendingReviewCount
      summary: Get pending review count
      description: Staff or admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending review count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PendingReviewCountResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/reviews/pending-by-product:
    get:
      tags: [Reviews]
      operationId: getPendingReviewsByProduct
      summary: Group pending reviews by product
      description: Staff or admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Pending review groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PendingReviewByProductResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/reviews/manage/product/{productId}:
    parameters:
      - $ref: '#/components/parameters/ProductIdPath'
    get:
      tags: [Reviews]
      operationId: listManagementReviews
      summary: List all reviews for moderation
      description: Staff or admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Approved and pending reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReviewResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/reviews/{reviewId}/reply:
    parameters:
      - $ref: '#/components/parameters/ReviewIdPath'
    post:
      tags: [Reviews]
      operationId: replyToReview
      summary: Reply to a review
      description: Staff or admin only. Review must already be approved.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReplyReviewRequest'
      responses:
        '200':
          description: Reply created or updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/reviews/{reviewId}/moderation:
    parameters:
      - $ref: '#/components/parameters/ReviewIdPath'
    put:
      tags: [Reviews]
      operationId: moderateReview
      summary: Approve or return a review to pending
      description: Staff or admin only.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModerateReviewRequest'
      responses:
        '200':
          description: Moderation applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewMutationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/reviews/{reviewId}:
    parameters:
      - $ref: '#/components/parameters/ReviewIdPath'
    delete:
      tags: [Reviews]
      operationId: deleteReview
      summary: Delete a review
      description: Staff or admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Review deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewMutationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/admin/statistics:
    get:
      tags: [Admin Statistics]
      operationId: getStatistics
      summary: Get summary statistics
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Summary metrics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatisticsSummaryResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/revenue:
    get:
      tags: [Admin Statistics]
      operationId: getRevenueDetails
      summary: Get revenue details
      description: Admin only.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Revenue series and totals
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevenueDetailsResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/revenue/export:
    get:
      tags: [Admin Statistics]
      operationId: exportRevenueCsv
      summary: Export revenue CSV
      description: Admin only.
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: from
          schema:
            type: string
            format: date
        - in: query
          name: to
          schema:
            type: string
            format: date
      responses:
        '200':
          description: CSV export
          content:
            text/csv:
              schema:
                type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ProductIdPath:
      in: path
      name: productId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    ReviewIdPath:
      in: path
      name: reviewId
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  responses:
    BadRequest:
      description: Request validation failed or payload is invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Authentication is required or credentials are invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated but not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Request conflicts with existing data or business rules
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
        error:
          type: string
        validationErrors:
          type: object
          additionalProperties:
            type: string
      required: [message, status, error]
...
    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          maxLength: 180
        brand:
          type: string
          maxLength: 120
        price:
          type: number
        imageUrl:
          type: string
          maxLength: 255
        categoryId:
          type: integer
          format: int64
        categoryName:
          type: string
        quantity:
          type: integer
        descriptionShort:
          type: string
        descriptionLong:
          type: string
      required: [id, name, price, quantity]
    ProductUpsertRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 180
        brand:
          type: string
          maxLength: 120
        price:
          type: number
          minimum: 0
        imageUrl:
          type: string
          maxLength: 255
        categoryId:
          type: integer
          format: int64
          minimum: 1
        quantity:
          type: integer
          minimum: 0
        descriptionShort:
          type: string
          maxLength: 2000
        descriptionLong:
          type: string
          maxLength: 10000
      required: [name, price, quantity]
...
    ReviewReplyResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        staffId:
          type: integer
          format: int64
        body:
          type: string
      required: [id, staffId, body]
    ReviewResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        userName:
          type: string
        userAvatar:
          type: string
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
        body:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
        reply:
          oneOf:
            - $ref: '#/components/schemas/ReviewReplyResponse'
            - type: 'null'
      required:
        - id
        - userId
        - userName
        - userAvatar
        - rating
        - title
        - body
        - isApproved
        - createdAt
        - reply
    PendingReviewCountResponse:
      type: object
      properties:
        pendingReviews:
          type: integer
      required: [pendingReviews]
    PendingReviewByProductResponse:
      type: object
      properties:
        productId:
          type: integer
          format: int64
        pendingReviews:
          type: integer
      required: [productId, pendingReviews]
    CreateReviewRequest:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title:
          type: string
          maxLength: 150
        body:
          type: string
          maxLength: 2000
      required: [rating, body]
    ReplyReviewRequest:
      type: object
      properties:
        body:
          type: string
          maxLength: 2000
      required: [body]
    ModerateReviewRequest:
      type: object
      properties:
        approved:
          type: boolean
      required: [approved]
    ReviewMutationResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        message:
          type: string
      required: [id, message]
...
    StatisticsSummaryResponse:
      type: object
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        totalUsers:
          type: integer
        totalProducts:
          type: integer
        deliveredOrders:
          type: integer
        deliveredRate:
          type: number
        avgOrderValue:
          type: number
        cancelRate:
          type: number
      required:
        - totalRevenue
        - totalOrders
        - totalUsers
        - totalProducts
        - deliveredOrders
        - deliveredRate
        - avgOrderValue
        - cancelRate
    RevenuePointByDateResponse:
      type: object
      properties:
        date:
          type: string
        revenue:
          type: number
      required: [date, revenue]
    RevenuePointByMonthResponse:
      type: object
      properties:
        month:
          type: string
        revenue:
          type: number
      required: [month, revenue]
    RevenueDetailsResponse:
      type: object
      properties:
        dayRevenue:
          type: number
        monthRevenue:
          type: number
        yearRevenue:
          type: number
        dailySeries:
          type: array
          items:
            $ref: '#/components/schemas/RevenuePointByDateResponse'
        monthlySeries:
          type: array
          items:
            $ref: '#/components/schemas/RevenuePointByMonthResponse'
      required: [dayRevenue, monthRevenue, yearRevenue, dailySeries, monthlySeries]